+++ react/README.md
@@ -4,11 +4,6 @@
 
 Приложение для самостоятельной работы в блоке React - веб-приложение погоды. На странице приложения должна быть возможность добавлять города в список избранных. По каждому городу показывается информация о температуре, ветре, другие параметры.
 
-    Развернуть приложение с помощью CRA (по желанию)
-    Настроить кастомный webpack + babel проект (+2 балла)
-    Заставить работать TypeScript через babel (+1 бал)
-    Написать хелло ворлд (todo app или любое другое), сделать базовую верстку вашего приложения, поработать с JSX (+2 бал)
-
 Создать структуру приложения, создать компоненты контейнеры.
 
 Источник данных прогноза
