+++ react/src/components/MyCitiesList/MyCitiesList.css
@@ -1,6 +1,7 @@
 .city-name {
     font-weight: bold;
     font-size: 1.2rem;
+    max-width: 130px;
 }
 
 .city-card-mini {
