+++ react/src/data/currentResponse.js
@@ -1,0 +1,41 @@
+export const currentResponse = {
+    "location": {
+        "name": "London",
+        "region": "City of London, Greater London",
+        "country": "United Kingdom",
+        "lat": 51.52,
+        "lon": -0.11,
+        "tz_id": "Europe/London",
+        "localtime_epoch": 1646563524,
+        "localtime": "2022-03-06 10:45"
+    },
+    "current": {
+        "last_updated_epoch": 1646562600,
+        "last_updated": "2022-03-06 10:30",
+        "temp_c": 6.0,
+        "temp_f": 42.8,
+        "is_day": 1,
+        "condition": {
+            "text": "Partly cloudy",
+            "icon": "//cdn.weatherapi.com/weather/64x64/day/116.png",
+            "code": 1003
+        },
+        "wind_mph": 12.5,
+        "wind_kph": 20.2,
+        "wind_degree": 70,
+        "wind_dir": "ENE",
+        "pressure_mb": 1029.0,
+        "pressure_in": 30.39,
+        "precip_mm": 0.0,
+        "precip_in": 0.0,
+        "humidity": 61,
+        "cloud": 75,
+        "feelslike_c": 2.5,
+        "feelslike_f": 36.6,
+        "vis_km": 10.0,
+        "vis_miles": 6.0,
+        "uv": 1.0,
+        "gust_mph": 14.5,
+        "gust_kph": 23.4
+    }
+};
