--- react/src/data/response.js
@@ -1,2970 +1,0 @@
-export const forecastResponse = {
-    "location": {
-        "name": "Moscow",
-        "region": "Moscow City",
-        "country": "Russia",
-        "lat": 55.75,
-        "lon": 37.62,
-        "tz_id": "Europe/Moscow",
-        "localtime_epoch": 1645902978,
-        "localtime": "2022-02-26 22:16"
-    },
-    "current": {
-        "last_updated_epoch": 1645902900,
-        "last_updated": "2022-02-26 22:15",
-        "temp_c": 2.0,
-        "temp_f": 35.6,
-        "is_day": 0,
-        "condition": {
-            "text": "Partly cloudy",
-            "icon": "//cdn.weatherapi.com/weather/64x64/night/116.png",
-            "code": 1003
-        },
-        "wind_mph": 6.9,
-        "wind_kph": 11.2,
-        "wind_degree": 330,
-        "wind_dir": "NNW",
-        "pressure_mb": 1023.0,
-        "pressure_in": 30.21,
-        "precip_mm": 0.0,
-        "precip_in": 0.0,
-        "humidity": 69,
-        "cloud": 75,
-        "feelslike_c": -0.7,
-    "feelslike_f": 30.8,
-    "vis_km": 10.0,
-    "vis_miles": 6.0,
-    "uv": 1.0,
-    "gust_mph": 10.3,
-    "gust_kph": 16.6
-},
-"forecast": {
-    "forecastday": [
-        {
-            "date": "2022-02-26",
-            "date_epoch": 1645833600,
-            "day": {
-                "maxtemp_c": 0.8,
-                "maxtemp_f": 33.4,
-                "mintemp_c": -5.1,
-                "mintemp_f": 22.8,
-                "avgtemp_c": -1.8,
-                "avgtemp_f": 28.7,
-                "maxwind_mph": 8.1,
-                "maxwind_kph": 13.0,
-                "totalprecip_mm": 0.1,
-                "totalprecip_in": 0.0,
-                "avgvis_km": 8.0,
-                "avgvis_miles": 4.0,
-                "avghumidity": 92.0,
-                "daily_will_it_rain": 0,
-                "daily_chance_of_rain": 0,
-                "daily_will_it_snow": 0,
-                "daily_chance_of_snow": 3,
-                "condition": {
-                    "text": "Overcast",
-                    "icon": "//cdn.weatherapi.com/weather/64x64/day/122.png",
-                    "code": 1009
-                },
-                "uv": 1.0
-            },
-            "astro": {
-                "sunrise": "07:29 AM",
-                "sunset": "05:57 PM",
-                "moonrise": "05:43 AM",
-                "moonset": "11:37 AM",
-                "moon_phase": "Waning Crescent",
-                "moon_illumination": "17"
-            },
-            "hour": [
-                {
-                    "time_epoch": 1645822800,
-                    "time": "2022-02-26 00:00",
-                    "temp_c": -4.8,
-                    "temp_f": 23.4,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 7.8,
-                    "wind_kph": 12.6,
-                    "wind_degree": 191,
-                    "wind_dir": "SSW",
-                    "pressure_mb": 1018.0,
-                    "pressure_in": 30.05,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 51,
-                    "feelslike_c": -9.8,
-                    "feelslike_f": 14.4,
-                    "windchill_c": -9.8,
-                    "windchill_f": 14.4,
-                    "heatindex_c": -4.8,
-                    "heatindex_f": 23.4,
-                    "dewpoint_c": -5.8,
-                    "dewpoint_f": 21.6,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 12.8,
-                    "gust_kph": 20.5,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645826400,
-                    "time": "2022-02-26 01:00",
-                    "temp_c": -5.1,
-                    "temp_f": 22.8,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/119.png",
-                        "code": 1006
-                    },
-                    "wind_mph": 7.4,
-                    "wind_kph": 11.9,
-                    "wind_degree": 192,
-                    "wind_dir": "SSW",
-                    "pressure_mb": 1017.0,
-                    "pressure_in": 30.04,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 71,
-                    "feelslike_c": -9.9,
-                    "feelslike_f": 14.2,
-                    "windchill_c": -9.9,
-                    "windchill_f": 14.2,
-                    "heatindex_c": -5.1,
-                    "heatindex_f": 22.8,
-                    "dewpoint_c": -6.0,
-                    "dewpoint_f": 21.2,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 12.3,
-                    "gust_kph": 19.8,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645830000,
-                    "time": "2022-02-26 02:00",
-                    "temp_c": -4.9,
-                    "temp_f": 23.2,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 7.2,
-                    "wind_kph": 11.5,
-                    "wind_degree": 197,
-                    "wind_dir": "SSW",
-                    "pressure_mb": 1017.0,
-                    "pressure_in": 30.04,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 100,
-                    "feelslike_c": -9.6,
-                    "feelslike_f": 14.7,
-                    "windchill_c": -9.6,
-                    "windchill_f": 14.7,
-                    "heatindex_c": -4.9,
-                    "heatindex_f": 23.2,
-                    "dewpoint_c": -5.8,
-                    "dewpoint_f": 21.6,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.6,
-                    "gust_kph": 18.7,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645833600,
-                    "time": "2022-02-26 03:00",
-                    "temp_c": -4.6,
-                    "temp_f": 23.7,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 6.9,
-                    "wind_kph": 11.2,
-                    "wind_degree": 203,
-                    "wind_dir": "SSW",
-                    "pressure_mb": 1017.0,
-                    "pressure_in": 30.03,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 100,
-                    "feelslike_c": -9.2,
-                    "feelslike_f": 15.4,
-                    "windchill_c": -9.2,
-                    "windchill_f": 15.4,
-                    "heatindex_c": -4.6,
-                    "heatindex_f": 23.7,
-                    "dewpoint_c": -5.5,
-                    "dewpoint_f": 22.1,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.2,
-                    "gust_kph": 18.0,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645837200,
-                    "time": "2022-02-26 04:00",
-                    "temp_c": -4.4,
-                    "temp_f": 24.1,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 6.3,
-                    "wind_kph": 10.1,
-                    "wind_degree": 209,
-                    "wind_dir": "SSW",
-                    "pressure_mb": 1017.0,
-                    "pressure_in": 30.04,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 100,
-                    "feelslike_c": -8.6,
-                    "feelslike_f": 16.5,
-                    "windchill_c": -8.6,
-                    "windchill_f": 16.5,
-                    "heatindex_c": -4.4,
-                    "heatindex_f": 24.1,
-                    "dewpoint_c": -5.2,
-                    "dewpoint_f": 22.6,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 10.1,
-                    "gust_kph": 16.2,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645840800,
-                    "time": "2022-02-26 05:00",
-                    "temp_c": -4.6,
-                    "temp_f": 23.7,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Light snow",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/326.png",
-                        "code": 1213
-                    },
-                    "wind_mph": 5.6,
-                    "wind_kph": 9.0,
-                    "wind_degree": 211,
-                    "wind_dir": "SSW",
-                    "pressure_mb": 1018.0,
-                    "pressure_in": 30.05,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 96,
-                    "cloud": 100,
-                    "feelslike_c": -8.5,
-                    "feelslike_f": 16.7,
-                    "windchill_c": -8.5,
-                    "windchill_f": 16.7,
-                    "heatindex_c": -4.6,
-                    "heatindex_f": 23.7,
-                    "dewpoint_c": -5.1,
-                    "dewpoint_f": 22.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 9.6,
-                    "gust_kph": 15.5,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645844400,
-                    "time": "2022-02-26 06:00",
-                    "temp_c": -4.2,
-                    "temp_f": 24.4,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Light snow",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/326.png",
-                        "code": 1213
-                    },
-                    "wind_mph": 5.1,
-                    "wind_kph": 8.3,
-                    "wind_degree": 218,
-                    "wind_dir": "SW",
-                    "pressure_mb": 1018.0,
-                    "pressure_in": 30.06,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 96,
-                    "cloud": 100,
-                    "feelslike_c": -7.8,
-                    "feelslike_f": 18.0,
-                    "windchill_c": -7.8,
-                    "windchill_f": 18.0,
-                    "heatindex_c": -4.2,
-                    "heatindex_f": 24.4,
-                    "dewpoint_c": -4.7,
-                    "dewpoint_f": 23.5,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 8.9,
-                    "gust_kph": 14.4,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645848000,
-                    "time": "2022-02-26 07:00",
-                    "temp_c": -3.5,
-                    "temp_f": 25.7,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 4.7,
-                    "wind_kph": 7.6,
-                    "wind_degree": 218,
-                    "wind_dir": "SW",
-                    "pressure_mb": 1019.0,
-                    "pressure_in": 30.08,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 96,
-                    "cloud": 90,
-                    "feelslike_c": -6.7,
-                    "feelslike_f": 19.9,
-                    "windchill_c": -6.7,
-                    "windchill_f": 19.9,
-                    "heatindex_c": -3.5,
-                    "heatindex_f": 25.7,
-                    "dewpoint_c": -4.0,
-                    "dewpoint_f": 24.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 8.3,
-                    "gust_kph": 13.3,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645851600,
-                    "time": "2022-02-26 08:00",
-                    "temp_c": -2.8,
-                    "temp_f": 27.0,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Light snow",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/326.png",
-                        "code": 1213
-                    },
-                    "wind_mph": 5.1,
-                    "wind_kph": 8.3,
-                    "wind_degree": 232,
-                    "wind_dir": "SW",
-                    "pressure_mb": 1019.0,
-                    "pressure_in": 30.09,
-                    "precip_mm": 0.1,
-                    "precip_in": 0.0,
-                    "humidity": 96,
-                    "cloud": 70,
-                    "feelslike_c": -6.1,
-                    "feelslike_f": 21.0,
-                    "windchill_c": -6.1,
-                    "windchill_f": 21.0,
-                    "heatindex_c": -2.8,
-                    "heatindex_f": 27.0,
-                    "dewpoint_c": -3.4,
-                    "dewpoint_f": 25.9,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 1,
-                    "chance_of_snow": 78,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 8.7,
-                    "gust_kph": 14.0,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645855200,
-                    "time": "2022-02-26 09:00",
-                    "temp_c": -1.1,
-                    "temp_f": 30.0,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 6.0,
-                    "wind_kph": 9.7,
-                    "wind_degree": 254,
-                    "wind_dir": "WSW",
-                    "pressure_mb": 1020.0,
-                    "pressure_in": 30.11,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 72,
-                    "feelslike_c": -4.5,
-                    "feelslike_f": 23.9,
-                    "windchill_c": -4.5,
-                    "windchill_f": 23.9,
-                    "heatindex_c": -1.1,
-                    "heatindex_f": 30.0,
-                    "dewpoint_c": -1.9,
-                    "dewpoint_f": 28.6,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 8.5,
-                    "gust_kph": 13.7,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645858800,
-                    "time": "2022-02-26 10:00",
-                    "temp_c": 0.0,
-                    "temp_f": 32.0,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 6.7,
-                    "wind_kph": 10.8,
-                    "wind_degree": 267,
-                    "wind_dir": "W",
-                    "pressure_mb": 1020.0,
-                    "pressure_in": 30.12,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 96,
-                    "cloud": 85,
-                    "feelslike_c": -3.5,
-                    "feelslike_f": 25.7,
-                    "windchill_c": -3.5,
-                    "windchill_f": 25.7,
-                    "heatindex_c": 0.0,
-                    "heatindex_f": 32.0,
-                    "dewpoint_c": -0.6,
-                    "dewpoint_f": 30.9,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 8.9,
-                    "gust_kph": 14.4,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645862400,
-                    "time": "2022-02-26 11:00",
-                    "temp_c": 0.6,
-                    "temp_f": 33.1,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 7.2,
-                    "wind_kph": 11.5,
-                    "wind_degree": 273,
-                    "wind_dir": "W",
-                    "pressure_mb": 1021.0,
-                    "pressure_in": 30.14,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 100,
-                    "feelslike_c": -3.0,
-                    "feelslike_f": 26.6,
-                    "windchill_c": -3.0,
-                    "windchill_f": 26.6,
-                    "heatindex_c": 0.6,
-                    "heatindex_f": 33.1,
-                    "dewpoint_c": -0.3,
-                    "dewpoint_f": 31.5,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 9.4,
-                    "gust_kph": 15.1,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645866000,
-                    "time": "2022-02-26 12:00",
-                    "temp_c": 0.8,
-                    "temp_f": 33.4,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 6.9,
-                    "wind_kph": 11.2,
-                    "wind_degree": 275,
-                    "wind_dir": "W",
-                    "pressure_mb": 1021.0,
-                    "pressure_in": 30.14,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 100,
-                    "feelslike_c": -2.6,
-                    "feelslike_f": 27.3,
-                    "windchill_c": -2.6,
-                    "windchill_f": 27.3,
-                    "heatindex_c": 0.8,
-                    "heatindex_f": 33.4,
-                    "dewpoint_c": -0.3,
-                    "dewpoint_f": 31.5,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 9.4,
-                    "gust_kph": 15.1,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645869600,
-                    "time": "2022-02-26 13:00",
-                    "temp_c": 0.7,
-                    "temp_f": 33.3,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 7.8,
-                    "wind_kph": 12.6,
-                    "wind_degree": 279,
-                    "wind_dir": "W",
-                    "pressure_mb": 1021.0,
-                    "pressure_in": 30.15,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 91,
-                    "cloud": 100,
-                    "feelslike_c": -3.1,
-                    "feelslike_f": 26.4,
-                    "windchill_c": -3.1,
-                    "windchill_f": 26.4,
-                    "heatindex_c": 0.7,
-                    "heatindex_f": 33.3,
-                    "dewpoint_c": -0.6,
-                    "dewpoint_f": 30.9,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 10.3,
-                    "gust_kph": 16.6,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645873200,
-                    "time": "2022-02-26 14:00",
-                    "temp_c": 0.2,
-                    "temp_f": 32.4,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 8.1,
-                    "wind_kph": 13.0,
-                    "wind_degree": 285,
-                    "wind_dir": "WNW",
-                    "pressure_mb": 1022.0,
-                    "pressure_in": 30.16,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 87,
-                    "cloud": 100,
-                    "feelslike_c": -3.8,
-                    "feelslike_f": 25.2,
-                    "windchill_c": -3.8,
-                    "windchill_f": 25.2,
-                    "heatindex_c": 0.2,
-                    "heatindex_f": 32.4,
-                    "dewpoint_c": -1.7,
-                    "dewpoint_f": 28.9,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.2,
-                    "gust_kph": 18.0,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645876800,
-                    "time": "2022-02-26 15:00",
-                    "temp_c": 0.1,
-                    "temp_f": 32.2,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 7.8,
-                    "wind_kph": 12.6,
-                    "wind_degree": 290,
-                    "wind_dir": "WNW",
-                    "pressure_mb": 1022.0,
-                    "pressure_in": 30.18,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 85,
-                    "cloud": 100,
-                    "feelslike_c": -3.8,
-                    "feelslike_f": 25.2,
-                    "windchill_c": -3.8,
-                    "windchill_f": 25.2,
-                    "heatindex_c": 0.1,
-                    "heatindex_f": 32.2,
-                    "dewpoint_c": -2.2,
-                    "dewpoint_f": 28.0,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.4,
-                    "gust_kph": 18.4,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645880400,
-                    "time": "2022-02-26 16:00",
-                    "temp_c": 0.4,
-                    "temp_f": 32.7,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 6.9,
-                    "wind_kph": 11.2,
-                    "wind_degree": 290,
-                    "wind_dir": "WNW",
-                    "pressure_mb": 1023.0,
-                    "pressure_in": 30.2,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 83,
-                    "cloud": 98,
-                    "feelslike_c": -3.1,
-                    "feelslike_f": 26.4,
-                    "windchill_c": -3.1,
-                    "windchill_f": 26.4,
-                    "heatindex_c": 0.4,
-                    "heatindex_f": 32.7,
-                    "dewpoint_c": -2.1,
-                    "dewpoint_f": 28.2,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 10.3,
-                    "gust_kph": 16.6,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645884000,
-                    "time": "2022-02-26 17:00",
-                    "temp_c": 0.4,
-                    "temp_f": 32.7,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 5.8,
-                    "wind_kph": 9.4,
-                    "wind_degree": 287,
-                    "wind_dir": "WNW",
-                    "pressure_mb": 1023.0,
-                    "pressure_in": 30.21,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 84,
-                    "cloud": 100,
-                    "feelslike_c": -2.7,
-                    "feelslike_f": 27.1,
-                    "windchill_c": -2.7,
-                    "windchill_f": 27.1,
-                    "heatindex_c": 0.4,
-                    "heatindex_f": 32.7,
-                    "dewpoint_c": -1.9,
-                    "dewpoint_f": 28.6,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 9.4,
-                    "gust_kph": 15.1,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645887600,
-                    "time": "2022-02-26 18:00",
-                    "temp_c": 0.2,
-                    "temp_f": 32.4,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 5.4,
-                    "wind_kph": 8.6,
-                    "wind_degree": 282,
-                    "wind_dir": "WNW",
-                    "pressure_mb": 1023.0,
-                    "pressure_in": 30.22,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 88,
-                    "cloud": 100,
-                    "feelslike_c": -2.7,
-                    "feelslike_f": 27.1,
-                    "windchill_c": -2.7,
-                    "windchill_f": 27.1,
-                    "heatindex_c": 0.2,
-                    "heatindex_f": 32.4,
-                    "dewpoint_c": -1.6,
-                    "dewpoint_f": 29.1,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 8.9,
-                    "gust_kph": 14.4,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645891200,
-                    "time": "2022-02-26 19:00",
-                    "temp_c": -0.6,
-                    "temp_f": 30.9,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 4.7,
-                    "wind_kph": 7.6,
-                    "wind_degree": 275,
-                    "wind_dir": "W",
-                    "pressure_mb": 1024.0,
-                    "pressure_in": 30.24,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 96,
-                    "feelslike_c": -3.3,
-                    "feelslike_f": 26.1,
-                    "windchill_c": -3.3,
-                    "windchill_f": 26.1,
-                    "heatindex_c": -0.6,
-                    "heatindex_f": 30.9,
-                    "dewpoint_c": -1.6,
-                    "dewpoint_f": 29.1,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 8.5,
-                    "gust_kph": 13.7,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645894800,
-                    "time": "2022-02-26 20:00",
-                    "temp_c": -2.2,
-                    "temp_f": 28.0,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 4.9,
-                    "wind_kph": 7.9,
-                    "wind_degree": 288,
-                    "wind_dir": "WNW",
-                    "pressure_mb": 1025.0,
-                    "pressure_in": 30.26,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 95,
-                    "cloud": 97,
-                    "feelslike_c": -5.3,
-                    "feelslike_f": 22.5,
-                    "windchill_c": -5.3,
-                    "windchill_f": 22.5,
-                    "heatindex_c": -2.2,
-                    "heatindex_f": 28.0,
-                    "dewpoint_c": -2.9,
-                    "dewpoint_f": 26.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 10.1,
-                    "gust_kph": 16.2,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645898400,
-                    "time": "2022-02-26 21:00",
-                    "temp_c": -2.3,
-                    "temp_f": 27.9,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 5.6,
-                    "wind_kph": 9.0,
-                    "wind_degree": 301,
-                    "wind_dir": "WNW",
-                    "pressure_mb": 1025.0,
-                    "pressure_in": 30.27,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 93,
-                    "feelslike_c": -5.8,
-                    "feelslike_f": 21.6,
-                    "windchill_c": -5.8,
-                    "windchill_f": 21.6,
-                    "heatindex_c": -2.3,
-                    "heatindex_f": 27.9,
-                    "dewpoint_c": -3.2,
-                    "dewpoint_f": 26.2,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 10.3,
-                    "gust_kph": 16.6,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645902000,
-                    "time": "2022-02-26 22:00",
-                    "temp_c": -1.5,
-                    "temp_f": 29.3,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 6.0,
-                    "wind_kph": 9.7,
-                    "wind_degree": 320,
-                    "wind_dir": "NW",
-                    "pressure_mb": 1026.0,
-                    "pressure_in": 30.28,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 96,
-                    "feelslike_c": -5.0,
-                    "feelslike_f": 23.0,
-                    "windchill_c": -5.0,
-                    "windchill_f": 23.0,
-                    "heatindex_c": -1.5,
-                    "heatindex_f": 29.3,
-                    "dewpoint_c": -2.4,
-                    "dewpoint_f": 27.7,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 9.6,
-                    "gust_kph": 15.5,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645905600,
-                    "time": "2022-02-26 23:00",
-                    "temp_c": -1.0,
-                    "temp_f": 30.2,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 6.0,
-                    "wind_kph": 9.7,
-                    "wind_degree": 331,
-                    "wind_dir": "NNW",
-                    "pressure_mb": 1026.0,
-                    "pressure_in": 30.3,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 95,
-                    "cloud": 100,
-                    "feelslike_c": -4.4,
-                    "feelslike_f": 24.1,
-                    "windchill_c": -4.4,
-                    "windchill_f": 24.1,
-                    "heatindex_c": -1.0,
-                    "heatindex_f": 30.2,
-                    "dewpoint_c": -1.8,
-                    "dewpoint_f": 28.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 9.2,
-                    "gust_kph": 14.8,
-                    "uv": 1.0
-                }
-            ]
-        },
-        {
-            "date": "2022-02-27",
-            "date_epoch": 1645920000,
-            "day": {
-                "maxtemp_c": -1.1,
-                "maxtemp_f": 30.0,
-                "mintemp_c": -7.0,
-                "mintemp_f": 19.4,
-                "avgtemp_c": -3.9,
-                "avgtemp_f": 24.9,
-                "maxwind_mph": 9.8,
-                "maxwind_kph": 15.8,
-                "totalprecip_mm": 0.1,
-                "totalprecip_in": 0.0,
-                "avgvis_km": 9.7,
-                "avgvis_miles": 6.0,
-                "avghumidity": 89.0,
-                "daily_will_it_rain": 0,
-                "daily_chance_of_rain": 0,
-                "daily_will_it_snow": 0,
-                "daily_chance_of_snow": 3,
-                "condition": {
-                    "text": "Partly cloudy",
-                    "icon": "//cdn.weatherapi.com/weather/64x64/day/116.png",
-                    "code": 1003
-                },
-                "uv": 1.0
-            },
-            "astro": {
-                "sunrise": "07:26 AM",
-                "sunset": "05:59 PM",
-                "moonrise": "06:38 AM",
-                "moonset": "12:52 PM",
-                "moon_phase": "Waning Crescent",
-                "moon_illumination": "9"
-            },
-            "hour": [
-                {
-                    "time_epoch": 1645909200,
-                    "time": "2022-02-27 00:00",
-                    "temp_c": -0.9,
-                    "temp_f": 30.4,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 6.5,
-                    "wind_kph": 10.4,
-                    "wind_degree": 340,
-                    "wind_dir": "NNW",
-                    "pressure_mb": 1027.0,
-                    "pressure_in": 30.31,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 100,
-                    "feelslike_c": -4.5,
-                    "feelslike_f": 23.9,
-                    "windchill_c": -4.5,
-                    "windchill_f": 23.9,
-                    "heatindex_c": -0.9,
-                    "heatindex_f": 30.4,
-                    "dewpoint_c": -1.8,
-                    "dewpoint_f": 28.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 10.1,
-                    "gust_kph": 16.2,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645912800,
-                    "time": "2022-02-27 01:00",
-                    "temp_c": -1.2,
-                    "temp_f": 29.8,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 7.4,
-                    "wind_kph": 11.9,
-                    "wind_degree": 346,
-                    "wind_dir": "NNW",
-                    "pressure_mb": 1027.0,
-                    "pressure_in": 30.33,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 92,
-                    "cloud": 100,
-                    "feelslike_c": -5.2,
-                    "feelslike_f": 22.6,
-                    "windchill_c": -5.2,
-                    "windchill_f": 22.6,
-                    "heatindex_c": -1.2,
-                    "heatindex_f": 29.8,
-                    "dewpoint_c": -2.4,
-                    "dewpoint_f": 27.7,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.2,
-                    "gust_kph": 18.0,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645916400,
-                    "time": "2022-02-27 02:00",
-                    "temp_c": -2.1,
-                    "temp_f": 28.2,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/119.png",
-                        "code": 1006
-                    },
-                    "wind_mph": 7.2,
-                    "wind_kph": 11.5,
-                    "wind_degree": 346,
-                    "wind_dir": "NNW",
-                    "pressure_mb": 1028.0,
-                    "pressure_in": 30.36,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 92,
-                    "cloud": 87,
-                    "feelslike_c": -6.2,
-                    "feelslike_f": 20.8,
-                    "windchill_c": -6.2,
-                    "windchill_f": 20.8,
-                    "heatindex_c": -2.1,
-                    "heatindex_f": 28.2,
-                    "dewpoint_c": -3.2,
-                    "dewpoint_f": 26.2,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.2,
-                    "gust_kph": 18.0,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645920000,
-                    "time": "2022-02-27 03:00",
-                    "temp_c": -3.0,
-                    "temp_f": 26.6,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 8.1,
-                    "wind_kph": 13.0,
-                    "wind_degree": 346,
-                    "wind_dir": "NNW",
-                    "pressure_mb": 1029.0,
-                    "pressure_in": 30.38,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 91,
-                    "feelslike_c": -7.7,
-                    "feelslike_f": 18.1,
-                    "windchill_c": -7.7,
-                    "windchill_f": 18.1,
-                    "heatindex_c": -3.0,
-                    "heatindex_f": 26.6,
-                    "dewpoint_c": -3.9,
-                    "dewpoint_f": 25.0,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 12.3,
-                    "gust_kph": 19.8,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645923600,
-                    "time": "2022-02-27 04:00",
-                    "temp_c": -4.0,
-                    "temp_f": 24.8,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 9.4,
-                    "wind_kph": 15.1,
-                    "wind_degree": 1,
-                    "wind_dir": "N",
-                    "pressure_mb": 1030.0,
-                    "pressure_in": 30.4,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 54,
-                    "feelslike_c": -9.4,
-                    "feelslike_f": 15.1,
-                    "windchill_c": -9.4,
-                    "windchill_f": 15.1,
-                    "heatindex_c": -4.0,
-                    "heatindex_f": 24.8,
-                    "dewpoint_c": -5.0,
-                    "dewpoint_f": 23.0,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 14.1,
-                    "gust_kph": 22.7,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645927200,
-                    "time": "2022-02-27 05:00",
-                    "temp_c": -5.9,
-                    "temp_f": 21.4,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 9.8,
-                    "wind_kph": 15.8,
-                    "wind_degree": 355,
-                    "wind_dir": "N",
-                    "pressure_mb": 1031.0,
-                    "pressure_in": 30.43,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 59,
-                    "feelslike_c": -11.9,
-                    "feelslike_f": 10.6,
-                    "windchill_c": -11.9,
-                    "windchill_f": 10.6,
-                    "heatindex_c": -5.9,
-                    "heatindex_f": 21.4,
-                    "dewpoint_c": -6.8,
-                    "dewpoint_f": 19.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 15.2,
-                    "gust_kph": 24.5,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645930800,
-                    "time": "2022-02-27 06:00",
-                    "temp_c": -6.9,
-                    "temp_f": 19.6,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/119.png",
-                        "code": 1006
-                    },
-                    "wind_mph": 8.9,
-                    "wind_kph": 14.4,
-                    "wind_degree": 359,
-                    "wind_dir": "N",
-                    "pressure_mb": 1031.0,
-                    "pressure_in": 30.45,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 64,
-                    "feelslike_c": -12.8,
-                    "feelslike_f": 9.0,
-                    "windchill_c": -12.8,
-                    "windchill_f": 9.0,
-                    "heatindex_c": -6.9,
-                    "heatindex_f": 19.6,
-                    "dewpoint_c": -7.8,
-                    "dewpoint_f": 18.0,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 14.1,
-                    "gust_kph": 22.7,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645934400,
-                    "time": "2022-02-27 07:00",
-                    "temp_c": -7.0,
-                    "temp_f": 19.4,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 9.4,
-                    "wind_kph": 15.1,
-                    "wind_degree": 359,
-                    "wind_dir": "N",
-                    "pressure_mb": 1032.0,
-                    "pressure_in": 30.48,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 61,
-                    "feelslike_c": -13.1,
-                    "feelslike_f": 8.4,
-                    "windchill_c": -13.1,
-                    "windchill_f": 8.4,
-                    "heatindex_c": -7.0,
-                    "heatindex_f": 19.4,
-                    "dewpoint_c": -7.9,
-                    "dewpoint_f": 17.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 14.5,
-                    "gust_kph": 23.4,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645938000,
-                    "time": "2022-02-27 08:00",
-                    "temp_c": -6.3,
-                    "temp_f": 20.7,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Light snow",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/326.png",
-                        "code": 1213
-                    },
-                    "wind_mph": 8.7,
-                    "wind_kph": 14.0,
-                    "wind_degree": 6,
-                    "wind_dir": "N",
-                    "pressure_mb": 1033.0,
-                    "pressure_in": 30.5,
-                    "precip_mm": 0.1,
-                    "precip_in": 0.0,
-                    "humidity": 91,
-                    "cloud": 65,
-                    "feelslike_c": -12.0,
-                    "feelslike_f": 10.4,
-                    "windchill_c": -12.0,
-                    "windchill_f": 10.4,
-                    "heatindex_c": -6.3,
-                    "heatindex_f": 20.7,
-                    "dewpoint_c": -7.5,
-                    "dewpoint_f": 18.5,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 1,
-                    "chance_of_snow": 83,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 12.3,
-                    "gust_kph": 19.8,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645941600,
-                    "time": "2022-02-27 09:00",
-                    "temp_c": -5.2,
-                    "temp_f": 22.6,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 8.7,
-                    "wind_kph": 14.0,
-                    "wind_degree": 11,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1034.0,
-                    "pressure_in": 30.52,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 88,
-                    "cloud": 100,
-                    "feelslike_c": -10.6,
-                    "feelslike_f": 12.9,
-                    "windchill_c": -10.6,
-                    "windchill_f": 12.9,
-                    "heatindex_c": -5.2,
-                    "heatindex_f": 22.6,
-                    "dewpoint_c": -6.9,
-                    "dewpoint_f": 19.6,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.4,
-                    "gust_kph": 18.4,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645945200,
-                    "time": "2022-02-27 10:00",
-                    "temp_c": -4.0,
-                    "temp_f": 24.8,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 8.5,
-                    "wind_kph": 13.7,
-                    "wind_degree": 10,
-                    "wind_dir": "N",
-                    "pressure_mb": 1034.0,
-                    "pressure_in": 30.53,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 83,
-                    "cloud": 100,
-                    "feelslike_c": -9.1,
-                    "feelslike_f": 15.6,
-                    "windchill_c": -9.1,
-                    "windchill_f": 15.6,
-                    "heatindex_c": -4.0,
-                    "heatindex_f": 24.8,
-                    "dewpoint_c": -6.4,
-                    "dewpoint_f": 20.5,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 10.7,
-                    "gust_kph": 17.3,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645948800,
-                    "time": "2022-02-27 11:00",
-                    "temp_c": -2.8,
-                    "temp_f": 27.0,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/119.png",
-                        "code": 1006
-                    },
-                    "wind_mph": 8.3,
-                    "wind_kph": 13.3,
-                    "wind_degree": 10,
-                    "wind_dir": "N",
-                    "pressure_mb": 1034.0,
-                    "pressure_in": 30.53,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 80,
-                    "cloud": 76,
-                    "feelslike_c": -7.5,
-                    "feelslike_f": 18.5,
-                    "windchill_c": -7.5,
-                    "windchill_f": 18.5,
-                    "heatindex_c": -2.8,
-                    "heatindex_f": 27.0,
-                    "dewpoint_c": -5.8,
-                    "dewpoint_f": 21.6,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 10.1,
-                    "gust_kph": 16.2,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645952400,
-                    "time": "2022-02-27 12:00",
-                    "temp_c": -1.8,
-                    "temp_f": 28.8,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Overcast",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/122.png",
-                        "code": 1009
-                    },
-                    "wind_mph": 8.3,
-                    "wind_kph": 13.3,
-                    "wind_degree": 8,
-                    "wind_dir": "N",
-                    "pressure_mb": 1034.0,
-                    "pressure_in": 30.53,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 77,
-                    "cloud": 95,
-                    "feelslike_c": -6.3,
-                    "feelslike_f": 20.7,
-                    "windchill_c": -6.3,
-                    "windchill_f": 20.7,
-                    "heatindex_c": -1.8,
-                    "heatindex_f": 28.8,
-                    "dewpoint_c": -5.2,
-                    "dewpoint_f": 22.6,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 10.1,
-                    "gust_kph": 16.2,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645956000,
-                    "time": "2022-02-27 13:00",
-                    "temp_c": -1.2,
-                    "temp_f": 29.8,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/119.png",
-                        "code": 1006
-                    },
-                    "wind_mph": 8.5,
-                    "wind_kph": 13.7,
-                    "wind_degree": 10,
-                    "wind_dir": "N",
-                    "pressure_mb": 1034.0,
-                    "pressure_in": 30.52,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 76,
-                    "cloud": 84,
-                    "feelslike_c": -5.6,
-                    "feelslike_f": 21.9,
-                    "windchill_c": -5.6,
-                    "windchill_f": 21.9,
-                    "heatindex_c": -1.2,
-                    "heatindex_f": 29.8,
-                    "dewpoint_c": -4.9,
-                    "dewpoint_f": 23.2,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 10.7,
-                    "gust_kph": 17.3,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645959600,
-                    "time": "2022-02-27 14:00",
-                    "temp_c": -1.1,
-                    "temp_f": 30.0,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 8.1,
-                    "wind_kph": 13.0,
-                    "wind_degree": 8,
-                    "wind_dir": "N",
-                    "pressure_mb": 1034.0,
-                    "pressure_in": 30.52,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 79,
-                    "cloud": 58,
-                    "feelslike_c": -5.3,
-                    "feelslike_f": 22.5,
-                    "windchill_c": -5.3,
-                    "windchill_f": 22.5,
-                    "heatindex_c": -1.1,
-                    "heatindex_f": 30.0,
-                    "dewpoint_c": -4.3,
-                    "dewpoint_f": 24.3,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.0,
-                    "gust_kph": 17.6,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1645963200,
-                    "time": "2022-02-27 15:00",
-                    "temp_c": -1.4,
-                    "temp_f": 29.5,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 8.7,
-                    "wind_kph": 14.0,
-                    "wind_degree": 5,
-                    "wind_dir": "N",
-                    "pressure_mb": 1034.0,
-                    "pressure_in": 30.53,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 81,
-                    "cloud": 46,
-                    "feelslike_c": -6.0,
-                    "feelslike_f": 21.2,
-                    "windchill_c": -6.0,
-                    "windchill_f": 21.2,
-                    "heatindex_c": -1.4,
-                    "heatindex_f": 29.5,
-                    "dewpoint_c": -4.3,
-                    "dewpoint_f": 24.3,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 12.3,
-                    "gust_kph": 19.8,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1645966800,
-                    "time": "2022-02-27 16:00",
-                    "temp_c": -2.4,
-                    "temp_f": 27.7,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 9.2,
-                    "wind_kph": 14.8,
-                    "wind_degree": 7,
-                    "wind_dir": "N",
-                    "pressure_mb": 1035.0,
-                    "pressure_in": 30.55,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 83,
-                    "cloud": 39,
-                    "feelslike_c": -7.3,
-                    "feelslike_f": 18.9,
-                    "windchill_c": -7.3,
-                    "windchill_f": 18.9,
-                    "heatindex_c": -2.4,
-                    "heatindex_f": 27.7,
-                    "dewpoint_c": -4.8,
-                    "dewpoint_f": 23.4,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 14.1,
-                    "gust_kph": 22.7,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1645970400,
-                    "time": "2022-02-27 17:00",
-                    "temp_c": -3.6,
-                    "temp_f": 25.5,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Sunny",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/113.png",
-                        "code": 1000
-                    },
-                    "wind_mph": 7.8,
-                    "wind_kph": 12.6,
-                    "wind_degree": 6,
-                    "wind_dir": "N",
-                    "pressure_mb": 1035.0,
-                    "pressure_in": 30.57,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 87,
-                    "cloud": 20,
-                    "feelslike_c": -8.3,
-                    "feelslike_f": 17.1,
-                    "windchill_c": -8.3,
-                    "windchill_f": 17.1,
-                    "heatindex_c": -3.6,
-                    "heatindex_f": 25.5,
-                    "dewpoint_c": -5.4,
-                    "dewpoint_f": 22.3,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 12.8,
-                    "gust_kph": 20.5,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1645974000,
-                    "time": "2022-02-27 18:00",
-                    "temp_c": -4.5,
-                    "temp_f": 23.9,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Clear",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/113.png",
-                        "code": 1000
-                    },
-                    "wind_mph": 8.1,
-                    "wind_kph": 13.0,
-                    "wind_degree": 4,
-                    "wind_dir": "N",
-                    "pressure_mb": 1035.0,
-                    "pressure_in": 30.58,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 90,
-                    "cloud": 21,
-                    "feelslike_c": -9.5,
-                    "feelslike_f": 14.9,
-                    "windchill_c": -9.5,
-                    "windchill_f": 14.9,
-                    "heatindex_c": -4.5,
-                    "heatindex_f": 23.9,
-                    "dewpoint_c": -5.8,
-                    "dewpoint_f": 21.6,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 13.2,
-                    "gust_kph": 21.2,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645977600,
-                    "time": "2022-02-27 19:00",
-                    "temp_c": -5.1,
-                    "temp_f": 22.8,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 8.3,
-                    "wind_kph": 13.3,
-                    "wind_degree": 9,
-                    "wind_dir": "N",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.58,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 91,
-                    "cloud": 54,
-                    "feelslike_c": -10.3,
-                    "feelslike_f": 13.5,
-                    "windchill_c": -10.3,
-                    "windchill_f": 13.5,
-                    "heatindex_c": -5.1,
-                    "heatindex_f": 22.8,
-                    "dewpoint_c": -6.3,
-                    "dewpoint_f": 20.7,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 13.6,
-                    "gust_kph": 22.0,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645981200,
-                    "time": "2022-02-27 20:00",
-                    "temp_c": -5.7,
-                    "temp_f": 21.7,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 8.1,
-                    "wind_kph": 13.0,
-                    "wind_degree": 14,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.59,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 42,
-                    "feelslike_c": -11.0,
-                    "feelslike_f": 12.2,
-                    "windchill_c": -11.0,
-                    "windchill_f": 12.2,
-                    "heatindex_c": -5.7,
-                    "heatindex_f": 21.7,
-                    "dewpoint_c": -6.7,
-                    "dewpoint_f": 19.9,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 13.9,
-                    "gust_kph": 22.3,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645984800,
-                    "time": "2022-02-27 21:00",
-                    "temp_c": -5.9,
-                    "temp_f": 21.4,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 8.1,
-                    "wind_kph": 13.0,
-                    "wind_degree": 22,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.6,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 50,
-                    "feelslike_c": -11.2,
-                    "feelslike_f": 11.8,
-                    "windchill_c": -11.2,
-                    "windchill_f": 11.8,
-                    "heatindex_c": -5.9,
-                    "heatindex_f": 21.4,
-                    "dewpoint_c": -6.8,
-                    "dewpoint_f": 19.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 13.4,
-                    "gust_kph": 21.6,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645988400,
-                    "time": "2022-02-27 22:00",
-                    "temp_c": -6.2,
-                    "temp_f": 20.8,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 7.6,
-                    "wind_kph": 12.2,
-                    "wind_degree": 21,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1037.0,
-                    "pressure_in": 30.61,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 55,
-                    "feelslike_c": -11.4,
-                    "feelslike_f": 11.5,
-                    "windchill_c": -11.4,
-                    "windchill_f": 11.5,
-                    "heatindex_c": -6.2,
-                    "heatindex_f": 20.8,
-                    "dewpoint_c": -7.0,
-                    "dewpoint_f": 19.4,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 12.8,
-                    "gust_kph": 20.5,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645992000,
-                    "time": "2022-02-27 23:00",
-                    "temp_c": -6.5,
-                    "temp_f": 20.3,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 7.2,
-                    "wind_kph": 11.5,
-                    "wind_degree": 17,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.6,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 59,
-                    "feelslike_c": -11.5,
-                    "feelslike_f": 11.3,
-                    "windchill_c": -11.5,
-                    "windchill_f": 11.3,
-                    "heatindex_c": -6.5,
-                    "heatindex_f": 20.3,
-                    "dewpoint_c": -7.2,
-                    "dewpoint_f": 19.0,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 12.1,
-                    "gust_kph": 19.4,
-                    "uv": 1.0
-                }
-            ]
-        },
-        {
-            "date": "2022-02-28",
-            "date_epoch": 1646006400,
-            "day": {
-                "maxtemp_c": -1.5,
-                "maxtemp_f": 29.3,
-                "mintemp_c": -9.1,
-                "mintemp_f": 15.6,
-                "avgtemp_c": -5.4,
-                "avgtemp_f": 22.3,
-                "maxwind_mph": 8.3,
-                "maxwind_kph": 13.3,
-                "totalprecip_mm": 0.2,
-                "totalprecip_in": 0.01,
-                "avgvis_km": 7.7,
-                "avgvis_miles": 4.0,
-                "avghumidity": 90.0,
-                "daily_will_it_rain": 0,
-                "daily_chance_of_rain": 65,
-                "daily_will_it_snow": 0,
-                "daily_chance_of_snow": 6,
-                "condition": {
-                    "text": "Light freezing rain",
-                    "icon": "//cdn.weatherapi.com/weather/64x64/day/311.png",
-                    "code": 1198
-                },
-                "uv": 1.0
-            },
-            "astro": {
-                "sunrise": "07:24 AM",
-                "sunset": "06:02 PM",
-                "moonrise": "07:15 AM",
-                "moonset": "02:22 PM",
-                "moon_phase": "Waning Crescent",
-                "moon_illumination": "2"
-            },
-            "hour": [
-                {
-                    "time_epoch": 1645995600,
-                    "time": "2022-02-28 00:00",
-                    "temp_c": -6.9,
-                    "temp_f": 19.6,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 7.4,
-                    "wind_kph": 11.9,
-                    "wind_degree": 16,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.6,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 95,
-                    "cloud": 65,
-                    "feelslike_c": -12.1,
-                    "feelslike_f": 10.2,
-                    "windchill_c": -12.1,
-                    "windchill_f": 10.2,
-                    "heatindex_c": -6.9,
-                    "heatindex_f": 19.6,
-                    "dewpoint_c": -7.6,
-                    "dewpoint_f": 18.3,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 12.8,
-                    "gust_kph": 20.5,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1645999200,
-                    "time": "2022-02-28 01:00",
-                    "temp_c": -7.2,
-                    "temp_f": 19.0,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 7.2,
-                    "wind_kph": 11.5,
-                    "wind_degree": 13,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.6,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 95,
-                    "cloud": 72,
-                    "feelslike_c": -12.4,
-                    "feelslike_f": 9.7,
-                    "windchill_c": -12.4,
-                    "windchill_f": 9.7,
-                    "heatindex_c": -7.2,
-                    "heatindex_f": 19.0,
-                    "dewpoint_c": -7.8,
-                    "dewpoint_f": 18.0,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 12.3,
-                    "gust_kph": 19.8,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646002800,
-                    "time": "2022-02-28 02:00",
-                    "temp_c": -7.4,
-                    "temp_f": 18.7,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 7.2,
-                    "wind_kph": 11.5,
-                    "wind_degree": 18,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1037.0,
-                    "pressure_in": 30.61,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 96,
-                    "cloud": 82,
-                    "feelslike_c": -12.6,
-                    "feelslike_f": 9.3,
-                    "windchill_c": -12.6,
-                    "windchill_f": 9.3,
-                    "heatindex_c": -7.4,
-                    "heatindex_f": 18.7,
-                    "dewpoint_c": -8.0,
-                    "dewpoint_f": 17.6,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 12.1,
-                    "gust_kph": 19.4,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646006400,
-                    "time": "2022-02-28 03:00",
-                    "temp_c": -7.4,
-                    "temp_f": 18.7,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 6.5,
-                    "wind_kph": 10.4,
-                    "wind_degree": 24,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1037.0,
-                    "pressure_in": 30.61,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 96,
-                    "cloud": 86,
-                    "feelslike_c": -12.3,
-                    "feelslike_f": 9.9,
-                    "windchill_c": -12.3,
-                    "windchill_f": 9.9,
-                    "heatindex_c": -7.4,
-                    "heatindex_f": 18.7,
-                    "dewpoint_c": -7.9,
-                    "dewpoint_f": 17.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 10.7,
-                    "gust_kph": 17.3,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646010000,
-                    "time": "2022-02-28 04:00",
-                    "temp_c": -7.3,
-                    "temp_f": 18.9,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Light snow",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/326.png",
-                        "code": 1213
-                    },
-                    "wind_mph": 6.7,
-                    "wind_kph": 10.8,
-                    "wind_degree": 21,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1037.0,
-                    "pressure_in": 30.61,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 96,
-                    "cloud": 84,
-                    "feelslike_c": -12.3,
-                    "feelslike_f": 9.9,
-                    "windchill_c": -12.3,
-                    "windchill_f": 9.9,
-                    "heatindex_c": -7.3,
-                    "heatindex_f": 18.9,
-                    "dewpoint_c": -7.9,
-                    "dewpoint_f": 17.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 10.5,
-                    "gust_kph": 16.9,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646013600,
-                    "time": "2022-02-28 05:00",
-                    "temp_c": -7.3,
-                    "temp_f": 18.9,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Light snow",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/326.png",
-                        "code": 1213
-                    },
-                    "wind_mph": 7.2,
-                    "wind_kph": 11.5,
-                    "wind_degree": 23,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1037.0,
-                    "pressure_in": 30.61,
-                    "precip_mm": 0.1,
-                    "precip_in": 0.0,
-                    "humidity": 95,
-                    "cloud": 80,
-                    "feelslike_c": -12.5,
-                    "feelslike_f": 9.5,
-                    "windchill_c": -12.5,
-                    "windchill_f": 9.5,
-                    "heatindex_c": -7.3,
-                    "heatindex_f": 18.9,
-                    "dewpoint_c": -8.0,
-                    "dewpoint_f": 17.6,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 1,
-                    "chance_of_snow": 84,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.2,
-                    "gust_kph": 18.0,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646017200,
-                    "time": "2022-02-28 06:00",
-                    "temp_c": -7.3,
-                    "temp_f": 18.9,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Light snow",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/326.png",
-                        "code": 1213
-                    },
-                    "wind_mph": 7.2,
-                    "wind_kph": 11.5,
-                    "wind_degree": 31,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1037.0,
-                    "pressure_in": 30.61,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 72,
-                    "feelslike_c": -12.5,
-                    "feelslike_f": 9.5,
-                    "windchill_c": -12.5,
-                    "windchill_f": 9.5,
-                    "heatindex_c": -7.3,
-                    "heatindex_f": 18.9,
-                    "dewpoint_c": -8.1,
-                    "dewpoint_f": 17.4,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.0,
-                    "gust_kph": 17.6,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646020800,
-                    "time": "2022-02-28 07:00",
-                    "temp_c": -6.9,
-                    "temp_f": 19.6,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 7.4,
-                    "wind_kph": 11.9,
-                    "wind_degree": 35,
-                    "wind_dir": "NE",
-                    "pressure_mb": 1037.0,
-                    "pressure_in": 30.61,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 91,
-                    "cloud": 56,
-                    "feelslike_c": -12.1,
-                    "feelslike_f": 10.2,
-                    "windchill_c": -12.1,
-                    "windchill_f": 10.2,
-                    "heatindex_c": -6.9,
-                    "heatindex_f": 19.6,
-                    "dewpoint_c": -8.1,
-                    "dewpoint_f": 17.4,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.0,
-                    "gust_kph": 17.6,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646024400,
-                    "time": "2022-02-28 08:00",
-                    "temp_c": -6.1,
-                    "temp_f": 21.0,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Patchy light snow",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/323.png",
-                        "code": 1210
-                    },
-                    "wind_mph": 8.3,
-                    "wind_kph": 13.3,
-                    "wind_degree": 36,
-                    "wind_dir": "NE",
-                    "pressure_mb": 1037.0,
-                    "pressure_in": 30.61,
-                    "precip_mm": 0.1,
-                    "precip_in": 0.0,
-                    "humidity": 89,
-                    "cloud": 77,
-                    "feelslike_c": -11.5,
-                    "feelslike_f": 11.3,
-                    "windchill_c": -11.5,
-                    "windchill_f": 11.3,
-                    "heatindex_c": -6.1,
-                    "heatindex_f": 21.0,
-                    "dewpoint_c": -7.6,
-                    "dewpoint_f": 18.3,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 69,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 11.2,
-                    "gust_kph": 18.0,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646028000,
-                    "time": "2022-02-28 09:00",
-                    "temp_c": -5.2,
-                    "temp_f": 22.6,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Sunny",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/113.png",
-                        "code": 1000
-                    },
-                    "wind_mph": 7.6,
-                    "wind_kph": 12.2,
-                    "wind_degree": 38,
-                    "wind_dir": "NE",
-                    "pressure_mb": 1037.0,
-                    "pressure_in": 30.62,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 87,
-                    "cloud": 24,
-                    "feelslike_c": -10.2,
-                    "feelslike_f": 13.6,
-                    "windchill_c": -10.2,
-                    "windchill_f": 13.6,
-                    "heatindex_c": -5.2,
-                    "heatindex_f": 22.6,
-                    "dewpoint_c": -7.0,
-                    "dewpoint_f": 19.4,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 9.8,
-                    "gust_kph": 15.8,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1646031600,
-                    "time": "2022-02-28 10:00",
-                    "temp_c": -4.4,
-                    "temp_f": 24.1,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 6.9,
-                    "wind_kph": 11.2,
-                    "wind_degree": 35,
-                    "wind_dir": "NE",
-                    "pressure_mb": 1037.0,
-                    "pressure_in": 30.63,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 84,
-                    "cloud": 62,
-                    "feelslike_c": -8.9,
-                    "feelslike_f": 16.0,
-                    "windchill_c": -8.9,
-                    "windchill_f": 16.0,
-                    "heatindex_c": -4.4,
-                    "heatindex_f": 24.1,
-                    "dewpoint_c": -6.7,
-                    "dewpoint_f": 19.9,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 8.7,
-                    "gust_kph": 14.0,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1646035200,
-                    "time": "2022-02-28 11:00",
-                    "temp_c": -3.2,
-                    "temp_f": 26.2,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 6.3,
-                    "wind_kph": 10.1,
-                    "wind_degree": 26,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1037.0,
-                    "pressure_in": 30.61,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 81,
-                    "cloud": 60,
-                    "feelslike_c": -7.2,
-                    "feelslike_f": 19.0,
-                    "windchill_c": -7.2,
-                    "windchill_f": 19.0,
-                    "heatindex_c": -3.2,
-                    "heatindex_f": 26.2,
-                    "dewpoint_c": -6.0,
-                    "dewpoint_f": 21.2,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 7.4,
-                    "gust_kph": 11.9,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1646038800,
-                    "time": "2022-02-28 12:00",
-                    "temp_c": -2.3,
-                    "temp_f": 27.9,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 6.7,
-                    "wind_kph": 10.8,
-                    "wind_degree": 13,
-                    "wind_dir": "NNE",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.6,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 79,
-                    "cloud": 55,
-                    "feelslike_c": -6.3,
-                    "feelslike_f": 20.7,
-                    "windchill_c": -6.3,
-                    "windchill_f": 20.7,
-                    "heatindex_c": -2.3,
-                    "heatindex_f": 27.9,
-                    "dewpoint_c": -5.5,
-                    "dewpoint_f": 22.1,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 7.8,
-                    "gust_kph": 12.6,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1646042400,
-                    "time": "2022-02-28 13:00",
-                    "temp_c": -1.6,
-                    "temp_f": 29.1,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 7.6,
-                    "wind_kph": 12.2,
-                    "wind_degree": 7,
-                    "wind_dir": "N",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.59,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 77,
-                    "cloud": 45,
-                    "feelslike_c": -5.8,
-                    "feelslike_f": 21.6,
-                    "windchill_c": -5.8,
-                    "windchill_f": 21.6,
-                    "heatindex_c": -1.6,
-                    "heatindex_f": 29.1,
-                    "dewpoint_c": -5.1,
-                    "dewpoint_f": 22.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 9.2,
-                    "gust_kph": 14.8,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1646046000,
-                    "time": "2022-02-28 14:00",
-                    "temp_c": -1.5,
-                    "temp_f": 29.3,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Partly cloudy",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/116.png",
-                        "code": 1003
-                    },
-                    "wind_mph": 7.4,
-                    "wind_kph": 11.9,
-                    "wind_degree": 7,
-                    "wind_dir": "N",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.58,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 78,
-                    "cloud": 36,
-                    "feelslike_c": -5.6,
-                    "feelslike_f": 21.9,
-                    "windchill_c": -5.6,
-                    "windchill_f": 21.9,
-                    "heatindex_c": -1.5,
-                    "heatindex_f": 29.3,
-                    "dewpoint_c": -4.8,
-                    "dewpoint_f": 23.4,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 9.4,
-                    "gust_kph": 15.1,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1646049600,
-                    "time": "2022-02-28 15:00",
-                    "temp_c": -1.6,
-                    "temp_f": 29.1,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Sunny",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/113.png",
-                        "code": 1000
-                    },
-                    "wind_mph": 6.5,
-                    "wind_kph": 10.4,
-                    "wind_degree": 8,
-                    "wind_dir": "N",
-                    "pressure_mb": 1035.0,
-                    "pressure_in": 30.57,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 81,
-                    "cloud": 18,
-                    "feelslike_c": -5.3,
-                    "feelslike_f": 22.5,
-                    "windchill_c": -5.3,
-                    "windchill_f": 22.5,
-                    "heatindex_c": -1.6,
-                    "heatindex_f": 29.1,
-                    "dewpoint_c": -4.5,
-                    "dewpoint_f": 23.9,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 8.5,
-                    "gust_kph": 13.7,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1646053200,
-                    "time": "2022-02-28 16:00",
-                    "temp_c": -1.9,
-                    "temp_f": 28.6,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Sunny",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/113.png",
-                        "code": 1000
-                    },
-                    "wind_mph": 5.4,
-                    "wind_kph": 8.6,
-                    "wind_degree": 0,
-                    "wind_dir": "N",
-                    "pressure_mb": 1035.0,
-                    "pressure_in": 30.57,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 85,
-                    "cloud": 6,
-                    "feelslike_c": -5.2,
-                    "feelslike_f": 22.6,
-                    "windchill_c": -5.2,
-                    "windchill_f": 22.6,
-                    "heatindex_c": -1.9,
-                    "heatindex_f": 28.6,
-                    "dewpoint_c": -4.0,
-                    "dewpoint_f": 24.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 7.2,
-                    "gust_kph": 11.5,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1646056800,
-                    "time": "2022-02-28 17:00",
-                    "temp_c": -2.9,
-                    "temp_f": 26.8,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Sunny",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/113.png",
-                        "code": 1000
-                    },
-                    "wind_mph": 4.0,
-                    "wind_kph": 6.5,
-                    "wind_degree": 346,
-                    "wind_dir": "NNW",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.58,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 92,
-                    "cloud": 0,
-                    "feelslike_c": -5.6,
-                    "feelslike_f": 21.9,
-                    "windchill_c": -5.6,
-                    "windchill_f": 21.9,
-                    "heatindex_c": -2.9,
-                    "heatindex_f": 26.8,
-                    "dewpoint_c": -4.0,
-                    "dewpoint_f": 24.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 5.6,
-                    "gust_kph": 9.0,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1646060400,
-                    "time": "2022-02-28 18:00",
-                    "temp_c": -3.7,
-                    "temp_f": 25.3,
-                    "is_day": 1,
-                    "condition": {
-                        "text": "Sunny",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/day/113.png",
-                        "code": 1000
-                    },
-                    "wind_mph": 4.5,
-                    "wind_kph": 7.2,
-                    "wind_degree": 348,
-                    "wind_dir": "NNW",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.58,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 93,
-                    "cloud": 0,
-                    "feelslike_c": -6.8,
-                    "feelslike_f": 19.8,
-                    "windchill_c": -6.8,
-                    "windchill_f": 19.8,
-                    "heatindex_c": -3.7,
-                    "heatindex_f": 25.3,
-                    "dewpoint_c": -4.6,
-                    "dewpoint_f": 23.7,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 6.5,
-                    "gust_kph": 10.4,
-                    "uv": 2.0
-                },
-                {
-                    "time_epoch": 1646064000,
-                    "time": "2022-02-28 19:00",
-                    "temp_c": -4.2,
-                    "temp_f": 24.4,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Clear",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/113.png",
-                        "code": 1000
-                    },
-                    "wind_mph": 4.5,
-                    "wind_kph": 7.2,
-                    "wind_degree": 354,
-                    "wind_dir": "N",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.59,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 0,
-                    "feelslike_c": -7.4,
-                    "feelslike_f": 18.7,
-                    "windchill_c": -7.4,
-                    "windchill_f": 18.7,
-                    "heatindex_c": -4.2,
-                    "heatindex_f": 24.4,
-                    "dewpoint_c": -5.1,
-                    "dewpoint_f": 22.8,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 6.5,
-                    "gust_kph": 10.4,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646067600,
-                    "time": "2022-02-28 20:00",
-                    "temp_c": -7.6,
-                    "temp_f": 18.3,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 4.7,
-                    "wind_kph": 7.6,
-                    "wind_degree": 356,
-                    "wind_dir": "N",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.6,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 2,
-                    "feelslike_c": -11.5,
-                    "feelslike_f": 11.3,
-                    "windchill_c": -11.5,
-                    "windchill_f": 11.3,
-                    "heatindex_c": -7.6,
-                    "heatindex_f": 18.3,
-                    "dewpoint_c": -8.4,
-                    "dewpoint_f": 16.9,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 9.8,
-                    "gust_kph": 15.8,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646071200,
-                    "time": "2022-02-28 21:00",
-                    "temp_c": -7.8,
-                    "temp_f": 18.0,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Clear",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/113.png",
-                        "code": 1000
-                    },
-                    "wind_mph": 4.3,
-                    "wind_kph": 6.8,
-                    "wind_degree": 6,
-                    "wind_dir": "N",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.6,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 94,
-                    "cloud": 3,
-                    "feelslike_c": -11.4,
-                    "feelslike_f": 11.5,
-                    "windchill_c": -11.4,
-                    "windchill_f": 11.5,
-                    "heatindex_c": -7.8,
-                    "heatindex_f": 18.0,
-                    "dewpoint_c": -8.6,
-                    "dewpoint_f": 16.5,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 10.0,
-                    "vis_miles": 6.0,
-                    "gust_mph": 8.9,
-                    "gust_kph": 14.4,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646074800,
-                    "time": "2022-02-28 22:00",
-                    "temp_c": -8.7,
-                    "temp_f": 16.3,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 2.5,
-                    "wind_kph": 4.0,
-                    "wind_degree": 358,
-                    "wind_dir": "N",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.6,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 96,
-                    "cloud": 3,
-                    "feelslike_c": -10.8,
-                    "feelslike_f": 12.6,
-                    "windchill_c": -10.8,
-                    "windchill_f": 12.6,
-                    "heatindex_c": -8.7,
-                    "heatindex_f": 16.3,
-                    "dewpoint_c": -9.2,
-                    "dewpoint_f": 15.4,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 5.1,
-                    "gust_kph": 8.3,
-                    "uv": 1.0
-                },
-                {
-                    "time_epoch": 1646078400,
-                    "time": "2022-02-28 23:00",
-                    "temp_c": -9.1,
-                    "temp_f": 15.6,
-                    "is_day": 0,
-                    "condition": {
-                        "text": "Mist",
-                        "icon": "//cdn.weatherapi.com/weather/64x64/night/143.png",
-                        "code": 1030
-                    },
-                    "wind_mph": 3.4,
-                    "wind_kph": 5.4,
-                    "wind_degree": 357,
-                    "wind_dir": "N",
-                    "pressure_mb": 1036.0,
-                    "pressure_in": 30.6,
-                    "precip_mm": 0.0,
-                    "precip_in": 0.0,
-                    "humidity": 96,
-                    "cloud": 5,
-                    "feelslike_c": -12.2,
-                    "feelslike_f": 10.0,
-                    "windchill_c": -12.2,
-                    "windchill_f": 10.0,
-                    "heatindex_c": -9.1,
-                    "heatindex_f": 15.6,
-                    "dewpoint_c": -9.6,
-                    "dewpoint_f": 14.7,
-                    "will_it_rain": 0,
-                    "chance_of_rain": 0,
-                    "will_it_snow": 0,
-                    "chance_of_snow": 0,
-                    "vis_km": 2.0,
-                    "vis_miles": 1.0,
-                    "gust_mph": 6.9,
-                    "gust_kph": 11.2,
-                    "uv": 1.0
-                }
-            ]
-        }
-    ]
-}
-}
