--- react/src/utils/ListWeatherUpdater.js
@@ -1,8 +1,0 @@
-class ListWeatherUpdater {
-    constructor() {}
-
-
-    update() {}
-}
-
-export default ListWeatherUpdater;
