+++ react/src/utils/city.states.js
@@ -1,0 +1,7 @@
+export const CITY_UNTOUCHED = 1;
+
+export const CITY_UPDATING = 2;
+
+export const CITY_DONE = 3;
+
+export const CITY_ERROR = 4;
